ci(perf_website_sale)

+++ perf_website_sale/models/product_tag.py
@@ -1,4 +1,9 @@
+import logging
+
 from odoo import models
+from odoo.modules.db import FunctionStatus
+
+_logger = logging.getLogger(__name__)
 
 
 class ProductTag(models.Model):
@@ -6,6 +11,12 @@ class ProductTag(models.Model):
 
     def init(self):
         res = super().init()
+        if self.env.registry.has_unaccent != FunctionStatus.INDEXABLE:
+            _logger.critical(
+                "UNACCENT is not present and INDEXABLE, cannot apply perf_website_sale"
+            )
+            return res
+
         indexes_to_remove = [
             "DROP INDEX IF EXISTS test_product_tag_visible_on_ecommerce",
             "DROP INDEX IF EXISTS test_product_tag_sequence",
